[Chore] Split payload.rs into six modules
1 341 lines covering seven unrelated jobs: dispatching a read object to
its writer, claiming an output path, the export manifest, where a
payload's public file goes, playable MonoBehaviour typetrees, image
encoding, and the multi-object container format.

    naming       where a payload's public file goes      (leaf)
    dedup        path claims, signatures, duplicates     (leaf)
    bundle       the multi-object container              (leaf)
    manifest     the per-export manifest                 -> bundle
    image_files  one image into every output format      -> bundle dedup
    playable     typetrees that carry a chart            -> dedup manifest naming

The facade keeps the dispatcher, 240 lines.

`playable` also calls back into the facade, and that is the design rather
than a leftover: a playable typetree writes its sub-payloads through the
same dispatcher any other object takes. The edge is written as
`use super::` with a comment, not hidden behind a glob.

Pure move: 47 items in, 47 out. A cycle check over the sibling edges
reports none.
