+++ src/core/export_pipeline/implementation/assetstudio.rs
@@ -26,9 +26,8 @@ use crate::core::config::{AppConfig, RegionConfig};
 use crate::core::errors::ExportPipelineError;
 
 use super::merge_phase_ms;
-use super::payload::{
-    is_playable_mono_typetree, write_assetstudio_playable_payloads, write_native_object_payload,
-};
+use super::payload::playable::{is_playable_mono_typetree, write_assetstudio_playable_payloads};
+use super::payload::write_native_object_payload;
 use super::selectors::{
     asset_studio_export_type_list, assetstudio_type_selector_matches,
     normalize_assetstudio_type_name,
